requirements.txt
numpy==1.26.4
pillow==10.3.0
opencv-python==4.9.0.80
glob2==0.7